Add more refactoring for comment components
